Fix code scanning alert - Multiplication result converted to larger type
Tracking issue for: https://github.com/poacpm/poac/security/code-scanning/1

+++ zutil.c
@@ -308,7 +308,7 @@ voidpf ZLIB_INTERNAL zcalloc (opaque, items, size)
     unsigned size;
 {
     (void)opaque;
-    return sizeof(uInt) > 2 ? (voidpf)malloc(items * size) :
+    return sizeof(uInt) > 2 ? (voidpf)malloc((unsigned long)items * size) :
                               (voidpf)calloc(items, size);
 }
